- structure refactor - docs_extractor.py updated
