+++ .gitignore
@@ -1,5 +1,6 @@
 .env
 config.py
+config/dev.py
 .ipynb_checkpoints
 __pycache__
 *.webm
