+++ config/__init__.py
@@ -1,4 +1,6 @@
-SECRET_KEY = ""
-MONGO_URI = ""
-# MONGO_URI = f"mongodb+srv://{username}:{password}@{host}/"
-# mongodb://[username:password@]host1[:port1][,...hostN[:portN]][/[defaultauthdb][?options]]
+environment = "dev"
+
+if environment == "dev":
+    from .dev import *
+else:
+    from .prod import *
