+++ config/prod.py
@@ -1,0 +1,4 @@
+SECRET_KEY = ""
+MONGO_URI = ""
+# MONGO_URI = f"mongodb+srv://{username}:{password}@{host}/"
+# mongodb://[username:password@]host1[:port1][,...hostN[:portN]][/[defaultauthdb][?options]]
