+++ knowledgegpt_base/utils/utils_docs.py
