+++ knowledgegpt_base/utils/utils_embedding.py
