+++ knowledgegpt_base/utils/utils_powerpoint.py
